limiting editing and deleting of board to owner

+++ controllers/userBoards.js
@@ -1,7 +1,7 @@
-const boardsRouter = require('express').Router()
+const userBoardsRouter = require('express').Router()
 const Board = require('../models/board')
 
-boardsRouter.get('/:userId', async (req, res) => {
+userBoardsRouter.get('/:userId', async (req, res) => {
   try {
     const boards = await Board.find({ members: req.params.userId })
     res.json(boards)
@@ -11,21 +11,7 @@ boardsRouter.get('/:userId', async (req, res) => {
   }
 })
 
-boardsRouter.put('/:boardId', async (req, res) => {
-  try {
-    const updatedBoard = await Board.findByIdAndUpdate(
-      req.params.boardId,
-      req.body,
-      { new: true }
-    )
-    res.json(updatedBoard)
-  } catch (error) {
-    console.log(error)
-    res.status(400).json({ error: 'something went wrong while editing board' })
-  }
-})
-
-boardsRouter.post('/', async (req, res) => {
+userBoardsRouter.post('/', async (req, res) => {
   try {
     const body = req.body
     const board = new Board({
@@ -42,11 +28,28 @@ boardsRouter.post('/', async (req, res) => {
   }
 })
 
-boardsRouter.delete('/:boardId', async (req, res) => {
+userBoardsRouter.put('/:boardId', async (req, res) => {
+  try {
+    const board = await Board.findById(req.params.boardId)
+    if(req.user.id !== board.owner.toString()) return res.status(401).end()
+    const updatedBoard = await Board.findByIdAndUpdate(
+      req.params.boardId,
+      req.body,
+      { new: true }
+    )
+    res.json(updatedBoard)
+  } catch (error) {
+    console.log(error)
+    res.status(400).json({ error: 'something went wrong while editing board' })
+  }
+})
+
+userBoardsRouter.delete('/:boardId', async (req, res) => {
   try {
     const board = await Board.findById(req.params.boardId).populate(
       'containers'
     )
+    if(req.user.id !== board.owner.toString()) return res.status(401).end()
     await board.containers.map(async container => {
       await container.cards.map(async card => {
         await Card.findByIdAndRemove(card)
@@ -61,4 +64,4 @@ boardsRouter.delete('/:boardId', async (req, res) => {
   }
 })
 
-module.exports = boardsRouter
+module.exports = userBoardsRouter
